+++ factory/random.py
@@ -1,8 +1,7 @@
 from __future__ import absolute_import
 
+import faker.generator
 import random
-
-from factory.faker import Faker
 
 randgen = random.Random()
 
@@ -17,7 +16,9 @@ def get_random_state():
 def set_random_state(state):
     """Force-set the state of factory.fuzzy's random generator."""
     randgen.state_set = True
-    return randgen.setstate(state)
+    randgen.setstate(state)
+
+    faker.generator.random.setstate(state)
 
 
 def reseed_random(seed):
@@ -25,6 +26,3 @@ def reseed_random(seed):
     r = random.Random(seed)
     random_internal_state = r.getstate()
     set_random_state(random_internal_state)
-
-    for locale in Faker._FAKER_REGISTRY:
-        Faker._FAKER_REGISTRY[locale].random.setstate(random_internal_state)