+++ docs/changelog.rst
@@ -8,6 +8,7 @@ ChangeLog
 
     - :issue:`451`: Restore :class:`~factory.django.FileField` to a
       :class:`~factory.declarations.ParameteredAttribute`, relying on composition to parse the provided parameters.
+    - :issue:`389`: Fix random state management with ``faker``.
 
 
 2.10.0 (2018-01-28)